Made ShortURL the namespace, moved classes into separate files.
