+++ lib/shorturl/exceptions.rb
@@ -1,0 +1,7 @@
+module ShortURL
+  class ServiceNotAvailable < Exception
+  end
+
+  class InvalidService < Exception
+  end
+end
